Move work details into the workMenuEntry component
